Day 28 complete!

+++ project/handmade.hpp
@@ -70,8 +70,8 @@ struct Player
 
 struct PlayerState
 {
-	s32 pos_x;
-	s32 pos_y;
+	f32 pos_x;
+	f32 pos_y;
 
 	b32 mid_jump;
 	f32 jump_time;
